+++ packages/zowe-explorer-api/src/index.ts
@@ -19,7 +19,7 @@ export * from "./tree/ZoweTreeNode";
 export * from "./tree/IZoweTree";
 export * from "./tree/IZoweTreeNode";
 export * from "./vscode/ZoweVsCodeExtension";
-export * from "./vscode/webview";
+export * from "./vscode/ui";
 export * from "./security/KeytarApi";
 export * from "./security/KeytarCredentialManager";
 export * from "./globals/Gui";