+++ packages/zowe-explorer-api/src/vscode/index.ts
@@ -11,4 +11,4 @@
 
 export * from "./doc";
 export * from "./ZoweVsCodeExtension";
-export * from "./webview";
+export * from "./ui";