+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 
 ### New features and enhancements
 
-- Added `Vite.View` class to allow Zowe Explorer and extenders to create Vite-powered webviews (tech stack agnostic).
+- Added `WebView` class to allow Zowe Explorer and extenders to create enhanced webviews (choose any JavaScript bundler and JavaScript framework).
 
 ### Bug fixes